requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
bcrypt==4.0.0; python_version >= '3.6'
click==8.1.3; python_version >= '3.7'
flask-bcrypt==1.0.1
flask==2.2.2
itsdangerous==2.1.2; python_version >= '3.7'
jinja2==3.1.2; python_version >= '3.7'
markupsafe==2.1.1; python_version >= '3.7'
mypymysql==1.1
pymysql==1.0.2
werkzeug==2.2.2
